[#1656] Refactor [client]
